fix(xNoiseSanplot): use analytical expression for Raylegh inverse cumulative distribution
